+++ packages/server-core/src/http/controllers/analysis-node-log/handlers/delete.ts
@@ -26,10 +26,6 @@ export async function deleteAnalysisNodeLogRouteHandler(req: Request, res: Respo
         },
     );
 
-    if (!isTelemetryClientUsable()) {
-        throw new BadRequestError('The telemetry service is not configured, therefore logs can not be deleted.');
-    }
-
     const filters : FiltersBuildInput<Log> = {
         labels: {
             entity: 'analysisNode',
@@ -50,10 +46,12 @@ export async function deleteAnalysisNodeLogRouteHandler(req: Request, res: Respo
 
     // todo: check permissions
 
-    const telemetryClient = useTelemetryClient();
-    await telemetryClient.log.deleteMany({
-        filters,
-    });
+    if (isTelemetryClientUsable()) {
+        const telemetryClient = useTelemetryClient();
+        await telemetryClient.log.deleteMany({
+            filters,
+        });
+    }
 
     return sendAccepted(res);
 }
